requirements.txt
APScheduler==3.10.4
arrow==1.2.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
schedule==1.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
zipp==3.16.2
